[Move] #3 CollectionGroup 소스코드 파일 통합
