--- KazPlaygroundSwiftUI/KazPlaygroundSwiftUI/Model/Review.swift
@@ -1,15 +1,0 @@
-//
-//  Review.swift
-//  KazPlaygroundSwiftUI
-//
-//  Created by 원태영 on 2023/01/20.
-//
-
-import Foundation
-
-struct Review : Identifiable {
-    let id : String
-    let storeID : String
-    let userID : String
-    let content : String
-}
