--- KazPlaygroundSwiftUI/KazPlaygroundSwiftUI/Model/Store.swift
@@ -1,13 +1,0 @@
-//
-//  Store.swift
-//  KazPlaygroundSwiftUI
-//
-//  Created by 원태영 on 2023/01/20.
-//
-
-import Foundation
-
-struct Store : Identifiable {
-    let id : String = UUID().uuidString
-    let name : String
-}
